[WIP] Improvements to modifier aliasing
Start of changes to modifier aliasing - when a modifer alias goes
down, if we detect a repeat, we can jump straight to the aliasing
state, since we know that repeat won't work for those keys, and a long
press and release probably doesn't mean "send the unaliased key once".

This commit is a start. We've updated a couple of tests and fixed up
the Armed state, but we still need to address MultiArmed, add
additional tests, and fix existing tests.